Xenon/requirements.txt
PyAutoGUI~=0.9.53
wikipedia~=1.4.0
art~=5.9
pyttsx3~=2.90
comtypes~=1.1.14
pycaw~=20230407
opencv-python~=4.7.0.72
Levenshtein~=0.20.9
openai~=0.27.2
Pillow~=9.5.0
requests~=2.28.2
psutil~=5.9.4
setuptools~=65.5.0
phonenumbers~=8.13.7
pyperclip~=1.8.2
pytube~=12.1.3
MouseInfo~=0.1.3